+++ runtime-common/stdlib/diagnostics/error-handling-functions.h
@@ -8,16 +8,16 @@
 #include <cstdint>
 
 #include "runtime-common/core/runtime-core.h"
-#include "runtime-common/stdlib/error-handling/error-handling-context.h"
+#include "runtime-common/stdlib/diagnostics/error-handling-context.h"
 
 inline int64_t f$error_reporting(int64_t level) noexcept {
   auto& error_handling_st{ErrorHandlingContext::get()};
-  const int32_t prev{error_handling_st.php_warning_level};
+  const int64_t prev{error_handling_st.php_warning_level};
   if ((level & E_ALL) == E_ALL) {
     error_handling_st.php_warning_level = php_warning_levels::PHP_WARNING_MAXIMUM_LEVEL;
   }
   if (php_warning_levels::PHP_WARNING_MINIMUM_LEVEL <= level && level <= php_warning_levels::PHP_WARNING_MAXIMUM_LEVEL) {
-    error_handling_st.php_warning_level = std::max(error_handling_st.php_warning_minimum_level, static_cast<int32_t>(level));
+    error_handling_st.php_warning_level = std::max(error_handling_st.php_warning_minimum_level, level);
   }
   return prev;
 }
